Seperated GUI and game Logic
