+++ guiWhy.java
@@ -18,40 +18,22 @@ import javafx.scene.text.*;
 import javafx.scene.image.*;
 
 
-// GAME LOGIC Libraries
-import java.util.Random; // roll dice
-import java.util.stream.Collectors;
-import java.util.Collections; // use collections methods such as frequency
-import java.util.ArrayList; // for arraylist 
-import java.util.List;  // for list
-
-// Libraries for writing and reading to file. Save and load game functionality
-import java.io.File; // for saving and loading game
-import java.io.FileWriter; // write files
-import java.io.IOException;
-//import java.util.Scanner; // reading from file
-
-
 public class guiWhy extends Application // Start of Class
 {
 
-    //Game Logic Vars
-    private int players=1; // number of players
-    private int currentPlayer = 0; // increment by one every main loop,  helps track current player
-    private Random rand = new Random(); // random for roll
-    private int reroll = 3; // tracks rerolls available
-    private int setButtonArr; // used to setOnAction for button array categories
-    Multiplayer[] playerobj; // holds each player obj, including their scoreboard, and overall score
+   
+    logic playerobj = new logic(); // holds each player obj, including their scoreboard, and overall score
    // int[] reservedDice = new int[5]; to be implemented later
+   private int setButtonArr; // used to setOnAction for button array categories
 
     // gui vars
     Stage window; // main windows that gui runs ins
     Scene scene1, scene2; // scene 1 is main menu and scene 2 will be main board screen
     Button [] categories = new Button [13]; // buttons to that score roll in a category
-    Label numberPlayers = new Label("Number of players: " + players); // label that contain's the amount of players playing
+    Label numberPlayers = new Label("Number of players: " + playerobj.getPlayers()); // label that contain's the amount of players playing
     Label rollVal = new Label("You rolled a: "); // value of the roll
-    Label curPlayer = new Label("Player "+(currentPlayer+1)+"'s Turn"); // who the current player is
-    Label numRolls = new Label("Rolls Left: "+ reroll); // number of rolls available per turn
+    Label curPlayer = new Label("Player "+(playerobj.getCurPlayer()+1)+"'s Turn"); // who the current player is
+    Label numRolls = new Label("Rolls Left: "+ playerobj.getreroll()); // number of rolls available per turn
     TextField input = new TextField("# of players?"); // field for the number of players playing
     private String [] options = new String[] {"Total of Aces: 1's","Total of Twos: 2's","Total of Threes: 3's", "Total of Fours: 4's",
     "Total of Fives: 5's", "Total of Sixes: 6's", "3 of A Kind | Total of All 5 Dice", "4 of A Kind | Total of All 5 Dice", "Full House | 25 points","Small Straight | 30 points", "Large Straight | 40 points", 
@@ -90,37 +72,37 @@ public class guiWhy extends Application // Start of Class
 
         Button startGame = new Button("Click me to play"); // button that when clicked, sets player (# of players) to textfield input and switches scenes
         Button instruct = new Button("Instructions");
-        startGame.setOnAction(e->setPlayer(input));
+        startGame.setOnAction(e->{ setPlayerGui(input);});
         input.setMaxWidth(310); 
-        input.setOnAction(e->setPlayer(input));
+        input.setOnAction(e->{ setPlayerGui(input);});
 
         //Layout 1 - elements in vertical column
         VBox layout1 = new VBox(20);
         layout1.setAlignment(Pos.BASELINE_CENTER);
         layout1.getChildren().addAll(label1, canvas, input, startGame, instruct);
-        scene1 = new Scene(layout1, 350, 300);
+        scene1 = new Scene(layout1, 350, 500);
 
         ///////////////////////////////////////////////////////////////////////////////////// START OF LAYOUT 2
 
         //Button 2
         Button Menu = new Button("Back to the Main Menu");
         Menu.setOnAction(e -> {
-          window.setScene(scene1); saveGame();}); // change scene to main menu and saveGame func
+          window.setScene(scene1); playerobj.saveGame();}); // change scene to main menu and saveGame func
 
         Button button3 = new Button("Next Player/Turn");
-        button3.setOnAction(e -> setCurrentPlayer()); // increment turn to next player or next roll if 1 player
+        button3.setOnAction(e -> {playerobj.setCurrentPlayer(); curLabel();}); // increment turn to next player or next roll if 1 player
 
         HBox topBorder = new HBox(); // HBox and Border Top
-        topBorder.setPadding(new Insets(10,0,-40,0));
+        topBorder.setPadding(new Insets(10,0,0,0));
         topBorder.setSpacing(10);
         curPlayer.setFont(Font.font("Arial Black",14));
-        HBox.setMargin(curPlayer,new Insets(5,0,-40,10));
+        HBox.setMargin(curPlayer,new Insets(5,0,0,0));
         topBorder.getChildren().addAll(Menu,button3,curPlayer);
 
         // Button Roll
         Button roll = new Button("Roll!"); // button to roll dice 
         roll.setFont(Font.font("Arial Black",14));
-        roll.setOnAction(e->YahtzeeRollLogic()); // roll 5 di, calls rollLogic function
+        roll.setOnAction(e->{rollSet();}); // roll 5 di, calls rollLogic function
 
         HBox bottomBorder = new HBox();
         bottomBorder.setSpacing(10);
@@ -138,7 +120,7 @@ public class guiWhy extends Application // Start of Class
         {
           int i = setButtonArr; // set i here because of final int issue if declared in for loop
           categories[setButtonArr] = new Button(options[i]);
-          categories[setButtonArr].setOnAction(e->yahtzeeScoreRoll(i+1)); // Score roll based on corresponding number
+          categories[setButtonArr].setOnAction(e->scoreRoll(i+1)); // Score roll based on corresponding number
           GridPane.setConstraints(categories[setButtonArr], 0,i); // add button at i to the layout
           categories[i].setMaxSize(Double.MAX_VALUE, Double.MAX_VALUE);
           categories[i].setFont(Font.font("Arial Black",14));
@@ -147,7 +129,7 @@ public class guiWhy extends Application // Start of Class
         layout2.setRight(eastBorder);
         layout2.setTop(topBorder);
         layout2.setBottom(bottomBorder);
-        scene2 = new Scene(layout2, 800, 675); // set scene to new scene
+        scene2 = new Scene(layout2, 800, 770); // set scene to new scene
 
         //Display scene 1 at first
         scene1.getStylesheets().add("style.css");
@@ -156,300 +138,61 @@ public class guiWhy extends Application // Start of Class
         window.show();
     }
 
-    public void setPlayer(TextField input)
-    {         // Try-catch that catches if entered value is not a valid integer
-        try            
-        {
-            players = Integer.parseInt(input.getText());    
-        } catch (NumberFormatException e) 
-        {
-            input.setText("Invalid value! Please enter a positive number");
-            return;
-        }
-            // Checks that passed in int is greater than 0s
-        if(Integer.parseInt(input.getText())<1 || Integer.parseInt(input.getText())>100)
-        return;
-
-        System.out.println(players);
-        YahtzeeSetupLogic();
-        numberPlayers.setText("Number of players: "+players);
+    public void setPlayerGui(TextField input)
+    {    
+      playerobj = new logic();
+      curPlayer.setText("Player "+(playerobj.getCurPlayer()+1)+"'s Turn");
+      if(playerobj.setPlayer(input))
+      {
+        numberPlayers.setText("Number of players: "+playerobj.getPlayers());
         window.setScene(scene2);
+      }
+      else return;
+    }
+
+    public void curLabel()
+    {
+       curPlayer.setText("Player "+(playerobj.getCurPlayer()+1)+"'s Turn");
+       System.out.print(playerobj.getCurPlayer());
+       numRolls.setText("Rolls Left: "+ playerobj.getreroll());
+       rollVal.setText("You rolled a: ");
+    }
+
+    public void rollSet()
+    {
+      if(playerobj.getreroll()!=0)
+      rollVal.setText("You rolled a: "); // reset roll label
+
+      boolean test = playerobj.YahtzeeRollLogic();
+      
+      if(!test)
+      {
+        numRolls.setText("OUT OF ROLLS!");
         return;
+      }
+
+      if(playerobj.getreroll()!=0)
+      numRolls.setText("Rolls Left: "+ playerobj.getreroll());
+
+      for(int i = 0; i < 5; i++) // set roll label to sorted roll
+        rollVal.setText(rollVal.getText()+ " " +playerobj.getMulti()[playerobj.getCurPlayer()].getRoll(i));
+ 
     }
 
-    public int getPlayer()
+    public void scoreRoll(int choice)
     {
-        return players;
-    }
-
-    public void setCurrentPlayer()
-    {
-        currentPlayer=(currentPlayer+1)%players;
-        curPlayer.setText("Player "+(currentPlayer+1)+"'s Turn");
-        System.out.print(currentPlayer);
-        reroll=3;
-        numRolls.setText("Rolls Left: "+ reroll);
-        rollVal.setText("You rolled a: ");
-    }
-
-    
-   public void YahtzeeSetupLogic()
-   {
-     System.out.print("Function started with given parameter "+ players + " ");
-     playerobj = new Multiplayer[players];
- 
-     for(int i =0; i<players; i++)
-     playerobj[i] = new Multiplayer(); // intialize player array
-   }
- 
-     // USED TO BE INFINITE LOOP
-   public void YahtzeeRollLogic()
-   {
-        if(reroll!=0)
-        rollVal.setText("You rolled a: "); // reset roll label
-
-         if(gameOver(playerobj))
-         {
-          System.out.print("GAME OVER");
-          return; // change to bool or kill statement
-         }
- 
-         if(fullScorecard(playerobj[currentPlayer]))
-         {
-           System.out.println("Player is out of categories to score!");
-           currentPlayer++;
-           return; 
-         }
-
-         if (reroll==0)
-         {
-         System.out.println("Player is out of rolls!");
-         return;
-         }
-        
-       System.out.println("\nPLAYER: "+ (currentPlayer+1) + "'s TURN");
-       reroll--; // rest reroll
-       if(reroll!=0)
-       numRolls.setText("Rolls Left: "+ reroll);
-       else
-       numRolls.setText("OUT OF ROLLS!");
-       
-       int rolled; // track single roll value
-        // roll loop that rolls 5 random dice, puts it in rolls, sorts rolls, and prints out roll. Also asks user for reroll
-       for(int i =0; i < 5; i++)
-       {
-        rolled = rand.nextInt(6)+1;
-        playerobj[currentPlayer].setRoll(i, rolled);
-       }
- 
-        playerobj[currentPlayer].sortRoll(); // sort in between print out for user readibility
-        for(int i = 0; i < 5; i++) // set roll label to sorted roll
-        rollVal.setText(rollVal.getText()+ " " +playerobj[currentPlayer].getRoll(i));
- 
-       for(int i =0; i < playerobj[currentPlayer].getRollSize(); i++) // print out roll
-         System.out.print("\t"+ playerobj[currentPlayer].getRoll(i));  
- 
-       System.out.println("\nREROLL?");
-       return;
-       }
-
-       
-       public void yahtzeeScoreRoll(int choice)
-       {
-        System.out.print(choice);
-        if(playerobj[currentPlayer].getScoreboard(choice -1)>-1 ) //  while choice is taken alert user and prompt for input
-       {
-         System.out.println("Option already taken or invalid" ); // + option picked by player AKA choice
-         //choice = cat.nextInt();
-         return;
-       }
-       if(rollVal.getText()=="You rolled a: ")
+           if(rollVal.getText()=="You rolled a: ")
        {
         System.out.println("You need to roll first!");
         return;
        }
- 
-       playerobj[currentPlayer].setScoreboard(choice-1,categories(playerobj[currentPlayer].getRollList(), choice)); // assign scoreboard category to score of dice based on result of category method
-       playerobj[currentPlayer].setScore(playerobj[currentPlayer].getScoreboard(choice-1)); // assign total score to itself + new points from category
- 
-       System.out.println("current score is:" + playerobj[currentPlayer].getScore()); // print our current score and score for that specific roll
-       System.out.println("Score for category " +choice+ " is " + playerobj[currentPlayer].getScoreboard(choice -1));
-       
-       setCurrentPlayer();
-     }
- 
-   
- 
-   public static int categories(ArrayList<Integer> rolls, int choice) // method to determine score of that roll in relation category chosen by user
-   {
-     int score = 0;
-    
-     if (choice < 7) // if less than 7, then score is the sum of dice that value was chosen by user (choice)
-     {
-         for(int i =0; i< rolls.size(); i++)
-         {
-         if(rolls.get(i)== choice)
-         score+=choice;
-         }
- 
-         return score;
-     }
-     
-     if (choice == 7 || choice == 8) // if 7 or 8 check frequency for three of a kind or four of a kind. If valid score is sum off all dice, otherwise 0
-     {
-         int threeKind = Collections.frequency(rolls, rolls.get(0));
-         int threeKindMid = Collections.frequency(rolls, rolls.get(2));
-         int threeKind2 = Collections.frequency(rolls, rolls.get(4));
-         if(choice==7)
-         {
-             if (threeKind > 2 || threeKind2 > 2 || threeKindMid > 2)
-             {
-                 for(int total : rolls)
-                     score+=total;
-             }
-         }
-         else if (choice == 8)
-         {
-             if (threeKind > 3 || threeKind2 > 3)
-             {
-                 for(int total : rolls)
-                     score+=total;
-             }
-         }
-         
-             return score;
-             
-     }
- 
-     if (choice == 9) // if 9, check for full house. Score is 25
-     {
-         int fullHouse = Collections.frequency(rolls, rolls.get(0));
-         int fullHouse2 = Collections.frequency(rolls, rolls.get(4));
- 
-          if  (  (fullHouse > 1 & fullHouse2 > 2) || (fullHouse > 2 && fullHouse2 > 1)  )
-             score+=25;
- 
-         return score;
-     }
- 
-     if (choice == 10) // if 10 look for small straight, 4 in a row. Score is 30
-     {
-       List<Integer> Straights = rolls.stream().distinct().collect(Collectors.toList());
-       if(Straights.size()<4) // convert rolls into distinct list with uniques integers
-       return score; // if size is less than 4, than can't be a small straight
- 
-       boolean straight=true;
- 
-         for(int i=0; i<3; i++)
-         {
-          if (Straights.get(i)!=Straights.get(i+1)-1)
-          straight=false;
-         }
-          if(straight==true)
-          return 30;
-       
-       //otherwise size is 5
-       if(Straights.size()==5)
-       {
-         for(int i=1; i<Straights.size()-1; i++)
-         {
-          if (Straights.get(i)!=Straights.get(i+1)-1)
-          straight=false;
-         }
-         if(straight==true)
-         return 30;
- 
-         for(int i=0; i<Straights.size()-2; i++)
-         { 
-          if (Straights.get(i)!=Straights.get(i+1)-1)
-          straight=false;
-        }
-         if(straight==true)
-         return 30;
-       }
-       return score;
-     }
-     if (choice == 11) // if 11 look for large straight, 5 in a row. Score is 40
-     {
-       boolean straight=true;
-       for(int i=0; i<rolls.size()-1; i++)
-       {
-         if (rolls.get(i)!=rolls.get(i+1)-1)
-         straight=false;
-       }
-       if(straight==true)
-       score += 40;
- 
-       return score;
-     }
- 
-     if (choice == 12) // if 12 look for five of a kind. Score is 50, YAHTZEE
-     {
-         int fiveKind = Collections.frequency(rolls, rolls.get(0));
-         if (fiveKind > 4)
-           score+=50;
- 
-         return score;
-     }
- 
-     if (choice == 13) // if 13 chance, score is sum of dice
-     {
-         for(int total : rolls)
-             score+=total;
- 
-             return score;
-     }
- 
-     return score;
-   }
- 
-   public static boolean fullScorecard (Multiplayer scoreBoard) // checks scoreboard to see if it is full
-   {
-       boolean full = true;
-       for (int i=0; i<13; i++)
-       {
-         if(scoreBoard.getScoreboard(i) <0)
-         full = false;
-       }
- 
-     return full;
-   }
- 
-   public static boolean gameOver (Multiplayer[] scoreBoard) // uses fullScorecard to check if every player's scorecard is full
-   {
-     boolean full = true;
-       for (int i=0; i<scoreBoard.length; i++)
-       {
-         if(!fullScorecard(scoreBoard[i]))
-         full = false;
-       }
- 
-     return full;
-   }
 
-   public void saveGame()
-   { // try block neccessarry due to ioexception 
-    try            
-    {
-      FileWriter writeSave = new FileWriter("saveGame.txt");
-      for(int i =0; i<13; i++)
-      {
-      writeSave.write("boo"); 
-      }
-
-
-      
-      writeSave.close();   
-    } catch (IOException e) 
-    {
-        input.setText("Issue with save file");
-        return;
+       if(playerobj.yahtzeeScoreRoll(choice))
+       {
+       rollVal.setText("You rolled a: ");
+       numRolls.setText("Rolls Left: "+ playerobj.getreroll());
+       curPlayer.setText("Player "+(playerobj.getCurPlayer()+1)+"'s Turn");
+       }
     }
-    
-   }
 
-   public void loadGame()
-   {
-    File saveFile = new File("saveGame.txt");
-   }
-}
+  }
