+++ style.css
@@ -8,5 +8,10 @@
 
 .label{
     -fx-text-fill: #000000;
+}
+
+.button{
+
+    -fx-text-fill: #000000;
     -fx-font-size: 20pt;
 }
